+++ server/controllers/schedule.controller.js
@@ -4,6 +4,7 @@ class ScheduleController {
     async create (req, res, next) {
         try {
             const { classId, schedule } = req.body;
+            if (!classId || !schedule) throw ApiError.BadRequest('Missing required fields');
             const createdSchedule = await scheduleService.create({classId, schedule});
             res.json(createdSchedule);
         } catch (e) {
@@ -14,6 +15,7 @@ class ScheduleController {
         try {
             const { scheduleId } = req.params;
             const { schedule } = req.body;
+            if (!scheduleId || !schedule) throw ApiError.BadRequest('Missing required fields');
             const updatedSchedule = await scheduleService.update({scheduleId, schedule});
             res.json(updatedSchedule);
         } catch (e) {
@@ -23,6 +25,7 @@ class ScheduleController {
     async delete (req, res, next) {
         try {
             const { scheduleId } = req.params;
+            if (!scheduleId) throw ApiError.BadRequest('Missing required fields');
             const deletedSchedule = await scheduleService.delete(scheduleId);
             res.json(deletedSchedule);
         } catch (e) {
